+++ ios/TikTokBusinessModule.mm
@@ -40,6 +40,7 @@ RCT_EXTERN_METHOD(initializeSdk:(NSString *)appId
                   ttAppId:(NSString *)ttAppId
                   accessToken:(NSString *)accessToken
                   debug:(nonnull NSNumber *)debug
+                  options:(NSDictionary *)options
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)